requirements.txt
attrs==25.3.0
Automat==25.4.16
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
constantly==23.10.4
cryptography==45.0.4
cssselect==1.3.0
defusedxml==0.7.1
filelock==3.18.0
Flask==3.1.1
greenlet==3.2.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
lxml==5.4.0
MarkupSafe==3.0.2
packaging==25.0
parsel==1.10.0
Protego==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==25.1.0
PySocks==1.7.1
python-dotenv==1.1.0
queuelib==1.8.0
requests==2.32.4
requests-file==2.1.0
Scrapy==2.13.2
service-identity==24.2.0
setuptools==80.9.0
soupsieve==2.7
SQLAlchemy==2.0.41
stem==1.8.2
tldextract==5.3.0
Twisted==25.5.0
typing_extensions==4.14.0
urllib3==2.4.0
w3lib==2.3.1
Werkzeug==3.1.3
zope.interface==7.2

# AI and Machine Learning Dependencies
google-generativeai>=0.3.0
spacy>=3.4.0

# OCR and Document Processing
pytesseract>=0.3.10
pillow>=9.0.0
pdf2image>=1.17.0

# Data Processing
pandas>=1.3.0
numpy>=1.21.0

# Additional utilities
colorama>=0.4.0
python-dateutil>=2.8.0
